refactor: opt interface
