+++ packages/design/src/shader-lab/IShaderInfo.ts
@@ -1,0 +1,6 @@
+import { ISubShaderInfo } from "./ISubShaderInfo";
+
+export interface IShaderInfo {
+  name: string;
+  subShaders: ISubShaderInfo[];
+}
