+++ packages/design/src/shader-lab/IShaderLab.ts
@@ -1,0 +1,8 @@
+import { IShaderInfo } from "./IShaderInfo";
+
+/**
+ * Shader lab interface.
+ */
+export interface IShaderLab {
+  parseShader(shaderSource: string): IShaderInfo;
+}
