+++ packages/design/src/shader-lab/IShaderPassInfo.ts
@@ -1,0 +1,7 @@
+export interface IShaderPassInfo {
+  name: string;
+  vert: string;
+  frag: string;
+  tags?: Record<string, number | string | boolean>;
+  renderStates: Record<string, any>;
+}
