+++ packages/design/src/shader-lab/ISubShaderInfo.ts
@@ -1,0 +1,6 @@
+import { IShaderPassInfo } from "./IShaderPassInfo";
+
+export interface ISubShaderInfo {
+  passes: IShaderPassInfo[];
+  tags?: Record<string, number | string | boolean>;
+}
