+++ packages/design/src/shader-lab/index.ts
@@ -1,1 +1,5 @@
-export * from "./interface";
+export type { IShaderInfo } from "./IShaderInfo";
+export type { IShaderLab } from "./IShaderLab";
+export type { IShaderPassInfo } from "./IShaderPassInfo";
+export type { ISubShaderInfo } from "./ISubShaderInfo";
+
