+++ packages/shader-lab/package.json
@@ -9,7 +9,6 @@
   "main": "dist/main.js",
   "module": "dist/module.js",
   "debug": "src/index.ts",
-  "browser": "dist/browser.js",
   "types": "types/index.d.ts",
   "scripts": {
     "b:types": "tsc",
